+++ src/components/detailViewComponents/subcomponents/CopyHashButton.vue
