+++ src/components/detailViewComponents/subcomponents/DetailHeader.vue
