+++ src/components/detailViewComponents/subcomponents/MetaDataPanel.vue
